\name{bilbao}
\docType{data}
\alias{bilbao}
\title{Bilbao waves data set}
\description{
  This data corresponds to the Bilbao waves data set, firstly analysed by Castillo and Hadi (1997) and in del Castillo and Serra (2015) from the MLE point of view.   
}
\usage{bilbao}
\format{A numeric vector.}
\references{
Castillo, E. and Hadi, A. S. (1997). Fitting the Generalized Pareto Distribution to Data. Journal of the American Statistical Association, \bold{92}, 1609-1620.
del Castillo, J. and Serra, I. (2015). Likelihood inference for Generalized Pareto Distribution. Computational Statistics and Data Analysis, \bold{83}, 116-128.
}
\keyword{datasets}