\name{d2}
\docType{data}
\alias{d2}
\title{EEMBC AutoBench suite (Benchmark 2)}
\description{
  This data corresponds to 1000 observations sampled from the second benchmark of the well-known suite for real-time systems EEMBC AutoBench suite (Poovey, 2007), including a number of of programs used in automotive embedded systems.   
}
\usage{d2}
\format{A numeric vector.}
\references{
Abella  J., Padilla, M.,del Castillo, J. & Cazorla, F. (2017). Measurement-Based Worst-Case Execution Time Estimation Using the Coefficient of Variation".ACM Transactions on Design Automation of Electronic Systems (TODAES), \bold{22}(4).

Poovey, J. (2007). Characterization of the EEMBC Benchmark Suite. North Carolina State University.
}
\keyword{datasets}