\encoding{utf8} 
\name{ercv-package}
\alias{ercv-package}
\alias{ercv}
\docType{package}
\title{
Empirical residual coefficient of variation
}
\description{
Fitting tails by the empirical residual coefficient of variation. 

}
\details{
\tabular{ll}{
Package: \tab ercv\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2019-09-19\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}
The package provides a methodology simple and trustworthy for the analysis of extreme values. The package contains functions for visualizing, fitting and validating the distribution of tails. Moreover, it also provides multiple threshold tests for a generalized Pareto distribution, together with an automatic threshold selection algorithm.
}

\author{
Joan del Castillo (Universitat Autònoma de Barcelona), David Moriña Soler (Catalan Institute of Oncology (ICO)-IDIBELL) and Isabel Serra (Centre de Recerca Matemàtica)
}

\seealso{
\code{\link{ercv-package}}, \code{\link{cievi}}, 
\code{\link{ccdfplot}}, \code{\link{cvevi}}, \code{\link{cvplot}}, \code{\link{evicv}}, \code{\link{fitpot}},
\code{\link{ppot}}, \code{\link{qpot}}, \code{\link{tdata}}, \code{\link{thrselect}}, 
\code{\link{Tm}}
}

\references{
del Castillo, J. and Padilla, M. (2016). Modeling extreme values by the residual coefficient of variation. SORT Statist. Oper. Res. Trans. \bold{40}(2), 303-320.

del Castillo, J. and Serra, I. (2015). Likelihood inference for Generalized Pareto Distribution. Computational Statistics and Data Analysis, \bold{83}, 116-128.

del Castillo, J., Daoudi, J. and Lockhart, R. (2014). Methods to Distinguish Between Polynomial and Exponential Tails. Scandinavian Journal of Statistics, \bold{41}, 382-393.
}
\keyword{ ercv }

