\name{EURUSD}
\docType{data}
\alias{EURUSD}
\title{Euro/Dollar daily exchange rates}
\description{
  This data corresponds to the euro/dollar daily exchange rates between 1999 and 2016, including the financial crisis of 2007-2008, which has been generated from the package \code{quantmod} (Ryan, 2016).   
}
\usage{EURUSD}
\format{A data frame with 6575 rows and 1 column.}
\references{
Ryan, J. A. (2016). quantmod: Quantitative Financial Modelling Framework. R package version 0.4-7. https://CRAN.R-project.org/package=quantmod
}
\keyword{datasets}